# CelebClothing/src/main/java/com/texstre/celebclothing/entity/DeliveryAddress.java
package com.texstre.celebclothing.entity;

import jakarta.persistence.*;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
@Table(name="delivery_address")
public class DeliveryAddress {
    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    @Column(name = "address_id")
    private Long addressId;
    @Column(name = "linked_usr_id")
    private Integer linkedUsrId;
    @Column(name = "contact_name")
    private String contactName;
    @Column(name = "contact_num")
    private String contactNum;
    @Column(name = "address_line1")
    private String addressLine1;
    @Column(name = "address_line2")
    private String addressLine2;
    @Column(name = "city")
    private String city;
    @Column(name = "state")
    private String state;
    @Column(name = "pin_code")
    private Integer pinCode;
    @Column(name = "country")
    private String country;
}