# CelebClothing/src/main/resources/texstreopenapi.yaml
# ROOT File
openapi: 3.1.1
info:
  title: TexStre E-commerce API
  description: |-
    An e-commerce API that allows various operations related to products, categories, and shopping carts.
  version: 1.0-SNAPSHOT
servers:
  - url: https://testenvironment.org/api/v1
  - url: https://prodenvironment.org/api/v1
tags:
  - name: products
    description: Operations related to products
  - name: categories
    description: Operations related to product categories
  - name: cart
    description: Operations related to shopping carts

paths:
  /users/{phoneNum}:
    get:
      summary: Get a user by Phone number
      operationId: fetchUserByPhoneNum
      parameters:
        - in: path
          name: phoneNum
          required: true
          schema:
            type: string
            minimum: 1
            description: The Phone or contact number identifier of the user.
      responses:
        "200":
          description: A single user with the given phone number.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /users:
    post:
      summary: Add a new User
      operationId: createUser
      requestBody:
        description: Create a new User using Phone number
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: Created

  /users/{userPhoneNum}:
    put:
      summary: Update details of existing User
      operationId: updateUser
      parameters:
        - in: path
          name: userPhoneNum
          required: true
          schema:
            type: string
            minimum: 1
            description: The Phone or contact number identifier of the user.
      requestBody:
        description: Update additional details of a User
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
      responses:
        "201":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

  /products:
    post:
      summary: Add a new Product
      operationId: createProduct
      requestBody:
        description: Create a new Product
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Product"
      responses:
        "201":
          description: Created

  /addresses:
    post:
      summary: Add a new Delivery Address
      operationId: createDeliveryAddress
      requestBody:
        description: Create a new Delivery Address
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeliveryAddress"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

  /addresses/{userId}:
    get:
      summary: Get a delivery address by userId
      operationId: fetchAddressesByUserId
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            minimum: 1
            description: The identifier of the user.
      responses:
        "200":
          description: A list of delivery addresses for the given userId
          content:
            application/json:
              schema:
                  $ref: "#/components/schemas/ApiResponse"

  /legal/tandc:
    get:
      summary: To provide Terms And Conditions as Response
      operationId: retrieveTermsAndConditions
      responses:
        "200":
          description: Terms and Conditions data from the file TermsAndConditions.txt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

  /legal/privacy_policy:
    get:
      summary: To provide Privacy Policy as Response
      operationId: retrievePrivacyPolicy
      responses:
        "200":
          description: Privacy policy data from the file PrivacyPolicy.txt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"

components:
  # Reusable schemas (data models)
  schemas:
    User:
      type: object
      properties:
        userId:
          type: integer
        contactNum:
          type: string
        email:
          type: string
        gender:
          type: string
        profilePicURL:
          type: string
        name:
          type: string
        userType:
          type: string

    Product:
      type: object
      properties:
        productName:
          type: string
        fabric:
          type: string
        basePrice:
          type: number
        sellPrice:
          type: number
        discPrice:
          type: number
        discPercent:
          type: integer
          format: int64

    DeliveryAddress:
      type: object
      properties:
        addressId:
          type: integer
        userId:
          type: integer
        contactName:
          type: string
        contactNum:
          type: string
        addressLine1:
          type: string
        addressLine2:
          type: string
        city:
          type: string
        state:
          type: string
        pinCode:
          type: integer
        country:
          type: string

    ApiResponse:
      type: object
      properties:
        http_code:
          type: integer
        message:
          type: string
        data:
          type: object
        metadata:
          type: object
  # Reusable path, query, header and cookie parameters
  #parameters: ...

  # Security scheme definitions (see Authentication)
  #securitySchemes: ...

  # Reusable request bodies
  #requestBodies: ...

  # Reusable responses, such as 401 Unauthorized or 400 Bad Request
  #responses: ...

  # Reusable response headers
  #headers: ...

  # Reusable examples
  #examples: ...

  # Reusable links
  #links: ...

  # Reusable callbacks
  #callbacks: